Removed --intelligent option
It was an incredibly convoluted solution to the problem which I was
trying to fix, and it simply didn't work. I do not want to try and
fix this at the moment as I believe --pad does the job well enough,
so I'm just gonna move all related code to a separate branch until
later.